bugfix, changed from int to double

+++ src/no/ntnu/ai/classifiers/nbc/NBCGenerator.java
@@ -14,17 +14,17 @@ public class NBCGenerator<T extends Number, T2> implements Generator<T, T2>{
 
 	@Override
 	public Hypothesis<T, T2> generateHypothesis(List<DataElement<T, T2>> data, Map<T2, Double> weights) {
-		Map<T2, Integer> occurences = new HashMap<T2, Integer>();
+		Map<T2, Double> occurences = new HashMap<T2, Double>();
 		for(DataElement<T, T2> dat : data){
 			if(!occurences.containsKey(dat.getClassification())){
-				occurences.put(dat.getClassification(), 0);
+				occurences.put(dat.getClassification(), 0.0);
 			}
 			occurences.put(dat.getClassification(), occurences.get(dat.getClassification()) + 1);
 		}
 		
 		Map<T2, Double> aProbs = new HashMap<T2, Double>();
 		for(T2 key : occurences.keySet()){
-			aProbs.put(key, (double)(occurences.get(key) / data.size()));
+			aProbs.put(key, (occurences.get(key) / data.size()));
 		}
 		
 		List<ProbGiven<T2>> attProbs = new ArrayList<ProbGiven<T2>>();
